third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.permission-default.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify Partitioned Popins in an iframe fails when the policy isn't set
  assert_equals: expected "Failure" but got "Success"
Harness: the test ran to completion.
